### 2020-04-16 #2177 @whedon
[REVIEW]: XLUR: A land use regression wizard for ArcGIS Pro
**Submitting author:** @anmolter (<a href="http://orcid.org/0000-0002-0824-4478">Anna Mölter</a>)
**Repository:** <a href="https://github.com/anmolter/XLUR" target ="_blank">https://github.com/anmolter/XLUR</a>
**Version:** v1.0.1
