### 2020-04-16 #2177 @arfon
@anmolter - this PR should fix your DOIs: https://github.com/anmolter/PyLUR/pull/1
