### 2020-04-25 #2177 @anmolter
@arfon  @hugoledoux  There was an article published in Frontiers of Environmental Science & Engineering last months (https://doi.org/10.1007/s11783-020-1221-5) about a GIS tool called 'PyLUR'. There's no link in the article to the tool, so I can't compare the code to mine, but reading the article I think there are definite differences, e.g. the published tool can't produce hybrid models, it doesn't have a wizard style interface etc.  However, I will need to rename my tool to avoid any confusion with the other PyLUR. Obviously, this won't change the source code just the documentation.
