### 2020-04-25 #2177 @arfon
> @arfon @hugoledoux Will renaming the repository cause any issues?

No, this should be fine. Just let us know the new repository address and the updated name for the package (you can update the `paper.md` yourself).
