+++ README.md
@@ -1,9 +1,9 @@
-# PyLUR
+# XLUR
 ## Usage
 A Python toolbox for ArcGIS Pro that enables the development and application of land use regression models.
 ## Instructions
-PyLUR was specifically developed as a toolbox for use within the ArcGIS Pro (v2.2.4 or higher, Environmental Systems Research Institute (ESRI), Redlands, CA) software. To use PyLUR the following steps are recommended:
-1. Clone or download the PyLUR repository to a local drive. Avoid having spaces in the file path to the directory where you are storing the PyLUR repository. 
-2. In the PyLUR repository open the PyLUR folder and doubleclick PyLUR_Manual_BrowserFriendly - Shortcut. This will open the PyLUR Wizard manual in a standard web browser. The manual contains the documentation for the PyLUR tools and a tutorial section, which guides the user through a test run of PyLUR. Alternatively, you can <a href="https://anmolter.github.io/PyLUR/PyLUR_documentation.pdf" type="application/pdf"> click here </a> to open a pdf version of the user manual in your web browser.
-3. Follow the instructions in the manual to install PyLUR and to launch the tools.  
+XLUR was specifically developed as a toolbox for use within the ArcGIS Pro (v2.2.4 or higher, Environmental Systems Research Institute (ESRI), Redlands, CA) software. To use XLUR the following steps are recommended:
+1. Clone or download the XLUR repository to a local drive. Avoid having spaces in the file path to the directory where you are storing the XLUR repository. 
+2. In the XLUR repository open the XLUR folder and doubleclick XLUR_Manual_BrowserFriendly - Shortcut. This will open the XLUR Wizard manual in a standard web browser. The manual contains the documentation for the XLUR tools and a tutorial section, which guides the user through a test run of XLUR. Alternatively, you can <a href="https://anmolter.github.io/PyLUR/PyLUR_documentation.pdf" type="application/pdf"> click here </a> to open a pdf version of the user manual in your web browser.
+3. Follow the instructions in the manual to install XLUR and to launch the tools.  
 
