+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'PyLUR: A land use regression wizard for ArcGIS Pro'
+title: 'XLUR: A land use regression wizard for ArcGIS Pro'
 tags:
   - Python
   - ArcGIS Pro
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 ## Introduction
 
-PyLUR is a Python toolbox for ArcGIS Pro (v2.2.4 or higher, Environmental Systems Research Institute (ESRI), Redlands, CA) 
+XLUR is a Python toolbox for ArcGIS Pro (v2.2.4 or higher, Environmental Systems Research Institute (ESRI), Redlands, CA) 
 that enables the development and application of land use regression models via a wizard style interface. 
 Land use regression (LUR) is a commonly used technique in environmental sciences to analyse factors influencing pollutant levels 
 and to predict pollutant levels at unmeasured locations. LUR is extensively used 
@@ -35,30 +35,30 @@ The traditional approach to the development of a LUR model requires several step
 4. Carrying out multiple regression analysis to obtain a best fit parsimonious model.
 
 If done manually, steps 2 to 4 are repetitive and time consuming, making them inefficient and prone to error. 
-PyLUR provides a wizard style interface that guides a user 
+XLUR provides a wizard style interface that guides a user 
 through the development of a LUR model without the need to access and run multiple tools and 
-additional software packages. PyLUR largely automates steps 2 to 4, which significantly speeds
+additional software packages. XLUR largely automates steps 2 to 4, which significantly speeds
 up the model development process and reduces user error. Depending on the type and amount of data to be processed, 
 and the available hardware, models can be developed in less than one hour. Furthermore, 
-the output files produced by PyLUR ensure that the model development is 
-well documented and reproducible. Lastly, PyLUR also makes the method available for a wider range of users.  
+the output files produced by XLUR ensure that the model development is 
+well documented and reproducible. Lastly, XLUR also makes the method available for a wider range of users.  
 
-In addition to developing LUR models, PyLUR can also apply a previously developed 
+In addition to developing LUR models, XLUR can also apply a previously developed 
 model to a new set of locations within the same study area. New locations can be
 defined by the user or can be based on a dataset containing regularly located points or randomly located points. 
-Again, PyLUR largely automates this process with minimal effort from the user. 
+Again, XLUR largely automates this process with minimal effort from the user. 
 
-PyLUR is aimed at GIS specialists. It uses the ArcGIS Pro software, which is the most widely 
-used commercial GIS software worldwide. PyLUR is based on the LUR methodology used in the 
+XLUR is aimed at GIS specialists. It uses the ArcGIS Pro software, which is the most widely 
+used commercial GIS software worldwide. XLUR is based on the LUR methodology used in the 
 European Study of Cohorts for Air Pollution Effects (ESCAPE) [@Beelen:2013;@Eeftens:2012] as set out in the ESCAPE Exposure 
 assessment manual [@escape]. Within air pollution research the ESCAPE methodology is used as
-the standard for developing LUR models. PyLUR also allows hybrid LUR models to be developed, based on an extension 
+the standard for developing LUR models. XLUR also allows hybrid LUR models to be developed, based on an extension 
 of the ESCAPE methodology that included the addition of satellite derived data and data from chemical transport 
 models [@deHoogh:2016]. 
 
 ## Current application
 
-PyLUR has been developed through the NERC Newton-DIPI funded Urban hybriD models for AiR pollution exposure Assessment (UDARA)
+XLUR has been developed through the NERC Newton-DIPI funded Urban hybriD models for AiR pollution exposure Assessment (UDARA)
[…]
