+++ XLUR/Code/apply.py
@@ -1,0 +1,2173 @@
+import wx
+import wx.html2
+import webbrowser
+from wx.lib.masked import NumCtrl
+from wx.lib.wordwrap import wordwrap
+from wx.lib.agw import ultimatelistctrl as ULC
+import wx.lib.inspection
+import arcpy
+from arcpy import env
+arcpy.CheckOutExtension("Spatial")
+import sys
+import os
+import logging
+import sqlite3
+import time
+import pandas as pd
+from pandas.tools.plotting import table
+import numpy as np
+import itertools
+import math
+import string
+import random
+import csv
+
+#------------------------------------------------------------------------------
+# User defined global functions
+#------------------------------------------------------------------------------
+# Function to capture system errors (Pokemon style) in file
+def exception_hook(exc_type, exc_value, exc_traceback):
+    logging.error( "Uncaught exception",exc_info=(exc_type, exc_value, exc_traceback))
+    arcpy.AddMessage('+++ERROR+++ Uncaught exception -> See GOTCHA')
+#------------------------------------------------------------------------------
+# Function to check that a field name exists
+def FieldExist(featureclass, fieldname):
+    fieldList = arcpy.ListFields(featureclass, fieldname)
+    fieldCount = len(fieldList)
+    if (fieldCount == 1):
+        return True
+    else:
+        return False
+#-------------------------------------------------------------------------------
+# Function to get unique values in field
+def unique_values(fc , field):
+    with arcpy.da.SearchCursor(fc, [field]) as cursor:
+        return {row[0] for row in cursor}
+#-------------------------------------------------------------------------------
+# Function to create a customised field map
+# This can be used to selectively import fields
+# Uses output field names in the dictionary
+def customFieldMap(in_fc,fldsNamesDict):
+    inputFieldNames=list(fldsNamesDict.keys())#list of input field names
+    customfieldmappings = arcpy.FieldMappings()# create an empty field mapping object
+    for field in inputFieldNames:# for each field, create an individual field map, and add it to the field mapping object
+        fmap = arcpy.FieldMap()
+        fmap.addInputField(in_fc, field)
+        outputFieldName = fmap.outputField
+        outputFieldName.name = fldsNamesDict[field]#use output field name from dictionary
+        outputFieldName.aliasName = fldsNamesDict[field]#aslo need to change alias name?
+        fmap.outputField = outputFieldName
+        customfieldmappings.addFieldMap(fmap)
+    return customfieldmappings
+#-------------------------------------------------------------------------------
+# load feature classes in SQLite DB, adapted from https://tereshenkov.wordpress.com/2016/12/17/load-esri-geodatabase-tables-into-sqlite/
+def run(SQL=None):
+    return conn.execute(SQL).fetchall()
+#------------------------------------------------------------------------------
+def fc_to_sql(conn,source_fc):
+    #getting info about the source feature class
+    source_basename = arcpy.Describe(source_fc).baseName
[…]
