+++ XLUR/Index/XLUR/log.txt
@@ -1,0 +1,1 @@
+Indexing D:\HDD\UDARA\LUR\Dev\ArcPro\XLUR\XLUR.aprx
