+++ PyLUR/Index/PyLUR/log.txt
@@ -1,1 +1,0 @@
-Indexing D:\HDD\UDARA\LUR\Dev\ArcPro\\PyLUR\PyLUR.aprx

[…]
