### 2020-04-26 #2177 @anmolter
@arfon  @hugoledoux  I've renamed the package and repository 'XLUR'. The address is: https://github.com/anmolter/XLUR . I've also updated paper.md and I've gone through the documentation and changed the name.
