### 2020-04-28 #2177 @fortinma
Hi @anmolter. I am stuck. I managed the install, but when I start the tutorial i get the following errors for pandas library, which is installed. See the screen cap for details.
  Any help in moving forward?
