### 2020-04-28 28dcccb
Added error catching for pandas.tool.plotting
