+++ XLUR/Code/build.py
@@ -14,7 +14,10 @@ import sqlite3
 import csv
 import time
 import pandas as pd
-from pandas.tools.plotting import table
+try:
+    from pandas.tools.plotting import table
+except ImportError:
+    from pandas.plotting import table
 import numpy as np
 import matplotlib as mpl
 import matplotlib.pyplot as plt
