### 2020-04-28 #2177 @anmolter
Hi @fortinma , I think you may have a newer version of pandas installed, where they changed pandas.tools. plotting to pandas.plotting. I've added  try except code with both versions. Is it working now?
