### 2020-04-28 #2177 @fortinma
hi again, @anmolter. The ApplyLUR script has the same issue with Pandas Unfortunately. See screencap.
![image](https://user-images.githubusercontent.com/1124016/80529653-c51dfb00-8965-11ea-8a3b-6aaff320cb6e.png)
