### 2020-04-29 d3d29c2
Added error checking for import of pandas.tools.plotting
