### 2020-04-29 #2177 @fortinma
Thanks, @anmolter. I just updated to ArcGIS Pro 2.5.0, which seems to be the latest -- for today at least.
