### 2020-04-30 #2177 @anmolter
@fortinma  Nice! The university I work for has disabled the automatic updates in ArcPro, so I'm stuck on version 2.2.4  😒
