### 2020-05-04 #2177 @anmolter
Hi @Athelena, could you close the wizard by clicking the x in the top right corner (don't click cancel, because that will probably delete the outputs), then go to your outputs folder and open the GOTCHA text file and let me know what that says?
