### 2020-05-04 #2177 @Athelena
`2020-05-04 13:23:25.407 ERROR build - exception_hook: Uncaught exception
Traceback (most recent call last):
  File "<string>", line 1, in <module>
AttributeError: 'ToolValidator' object has no attribute 'isLicensed'
2020-05-04 13:23:26.085 ERROR build - exception_hook: Uncaught exception
Traceback (most recent call last):
  File "<string>", line 1, in <module>
AttributeError: 'ToolValidator' object has no attribute 'isLicensed'
2020-05-04 13:27:34.045 ERROR build - exception_hook: Uncaught exception
Traceback (most recent call last):
  File "<string>", line 1, in <module>
AttributeError: 'ToolValidator' object has no attribute 'isLicensed'
2020-05-04 13:27:34.541 ERROR build - exception_hook: Uncaught exception
Traceback (most recent call last):
  File "<string>", line 1, in <module>
AttributeError: 'ToolValidator' object has no attribute 'isLicensed'
2020-05-04 13:31:23.324 ERROR build - exception_hook: Uncaught exception
Traceback (most recent call last):
  File "<string>", line 1, in <module>
AttributeError: 'ToolValidator' object has no attribute 'isLicensed'
2020-05-04 13:31:23.789 ERROR build - exception_hook: Uncaught exception
Traceback (most recent call last):
  File "<string>", line 1, in <module>
AttributeError: 'ToolValidator' object has no attribute 'isLicensed'
2020-05-04 13:39:29.410 ERROR build - exception_hook: Uncaught exception
Traceback (most recent call last):
  File "\\Mac\Home\Downloads\XLUR-master\XLUR\Code\build.py", line 5853, in onNext
    arcpy.sa.ExtractValuesToPoints(out_fds+"\\sites",out_fgdb+"\\"+pG_name,out_fgdb+"\\"+pG_name+"_sites","","VALUE_ONLY") #join cell values to points
  File "c:\users\alabetski\appdata\local\programs\arcgis\pro\Resources\arcpy\arcpy\sa\Functions.py", line 1645, in ExtractValuesToPoints
    add_attributes)
  File "c:\users\alabetski\appdata\local\programs\arcgis\pro\Resources\arcpy\arcpy\sa\Utils.py", line 53, in swapper
    result = wrapper(*args, **kwargs)
  File "c:\users\alabetski\appdata\local\programs\arcgis\pro\Resources\arcpy\arcpy\sa\Functions.py", line 1638, in Wrapper
    add_attributes)
  File "c:\users\alabetski\appdata\local\programs\arcgis\pro\Resources\arcpy\arcpy\geoprocessing\_base.py", line 511, in <lambda>
    return lambda *args: val(*gp_fixargs(args, True))
arcgisscripting.ExecuteError: Failed to execute. Parameters are not valid.
ERROR 000824: The tool is not licensed.
Failed to execute (ExtractValuesToPoints).
`
