### 2020-05-04 #2177 @anmolter
@Athelena It looks like you don't have a license for Spatial Analyst. Could you check your Licensing information? Here's a link  for more info: https://pro.arcgis.com/en/pro-app/help/analysis/spatial-analyst/basics/enabling-the-spatial-analyst-extension.htm
