### 2020-05-05 6ddc037
Added requirement for Spatial Analyst license if raster analysis, fixed bug in apply to skip predictor types, added workspace reset in p3G
