+++ XLUR/Code/apply.py
@@ -7,7 +7,6 @@ from wx.lib.agw import ultimatelistctrl as ULC
 import wx.lib.inspection
 import arcpy
 from arcpy import env
-arcpy.CheckOutExtension("Spatial")
 import sys
 import os
 import logging
@@ -418,7 +417,16 @@ class WizardPanel1(wx.Panel):
             p1and2.append(p[:p.find('_',3)])
 
         p1and2 = list(set(p1and2)) # only keep unique ones
-        #check that files exsist and for A,B,C that cat field exists
+        #check if raster data and check out extension, warning if no license
+        for item in p1and2:
+            if item[0:2]=='pG':
+                try:
+                    arcpy.CheckOutExtension("Spatial")
+                except:
+                    wx.MessageBox('The Spatial Analyst extension could not be activated. Please check your licenses or contact your system administrator.','Error',wx.OK|wx.ICON_ERROR)
+                    del wait
+
+        #check that files exist and for A,B,C that cat field exists
         for item in p1and2:
             if not arcpy.Exists(lur_fgdb+"\\"+item):
                 log.write('\n +++ERROR+++ '+item+' not found in LUR file geodatabase.')
@@ -1581,423 +1589,427 @@ class WizardPanel3(wx.Panel):
         conn.execute('''INSERT INTO timings VALUES('apply_panel3_ABC','start',datetime(),NULL)''')
         #Buffer+Intersect predictors
         preds_ABC = [i for i in preds if i[:2] in ['pA','pB','pC']]
-        # extract first two parts and filter unique
-        preds_ABC_1to2 = [i[:i.find('_',3)] for i in preds_ABC]
-        preds_ABC_1to2 = list(set(preds_ABC_1to2)) #these are the fc that need to be analysed
-        #extract buffer distances and run multiple ring buffer
-        p4 = [int(i[i.rfind('_',0,len(i)-4)+1:-4]) for i in preds_ABC] # get buffer distances
-        p4 = list(set(p4)) # only unique values
-        arcpy.MultipleRingBuffer_analysis(out_recp, out_fds+"\\MultBuffer", p4, "", "", "NONE") # buffer
-        arcpy.AddMessage('\nbuffer created')
+        if len(preds_ABC)>0:#check that there are ABC predictors
+            # extract first two parts and filter unique
+            preds_ABC_1to2 = [i[:i.find('_',3)] for i in preds_ABC]
+            preds_ABC_1to2 = list(set(preds_ABC_1to2)) #these are the fc that need to be analysed
+            #extract buffer distances and run multiple ring buffer
+            p4 = [int(i[i.rfind('_',0,len(i)-4)+1:-4]) for i in preds_ABC] # get buffer distances
+            p4 = list(set(p4)) # only unique values
+            arcpy.MultipleRingBuffer_analysis(out_recp, out_fds+"\\MultBuffer", p4, "", "", "NONE") # buffer
+            arcpy.AddMessage('\nbuffer created')
 
-        #iterate through feature classes
-        for item in preds_ABC_1to2:
-            arcpy.AddMessage(("\nStarting: "+item))
-            temp_preds_ABC = [i for i in preds_ABC if i.startswith(item+'_')] # get predictor variables belonging to item
-            temp_p4 = [int(i[i.rfind('_',0,len(i)-4)+1:-4]) for i in temp_preds_ABC] # get buffer distances from predictors
-            temp_p4 = list(set(temp_p4)) # only keep unique values
-            temp_p3 = [i[i.find('_',3)+1:i.rfind('_',0,len(i)-4)] for i in temp_preds_ABC] # get categories from predictors
-            temp_p3 = list(set(temp_p3)) # only keep unique values
[…]
