+++ XLUR/Code/Documentation/XLUR_manual_BrowserFriendly.Rmd
@@ -1,0 +1,1313 @@
+---
+title: XLUR Wizard
+date: "v.`r Sys.Date()`"
+output:
+  rmdformats::material:
+    highlight: kate
+    self_contained: no
+    lightbox: TRUE
+    fig_caption: FALSE
+
+---
+
+
+
+```{r knitr_init, echo=FALSE, cache=FALSE}
+library(knitr)
+library(rmdformats)
+
+## Global options
+options(max.print="75")
+opts_chunk$set(echo=FALSE,
+	             cache=TRUE,
+               prompt=FALSE,
+               tidy=TRUE,
+               comment=NA,
+               message=FALSE,
+               warning=FALSE)
+opts_knit$set(width=75)
+```
+
+# Installation
+
+System requirements: Windows 10, ArcGIS Pro v2.2.4 or higher, License for Spatial Analyst extension if analysis of raster data is required
+
+The XLUR tools require a small number of additional Python packages to be installed. In ArcGIS Pro this can be done using the Python Package Manager. Follow these steps to install additional Python packages:
+
+1. Open ArcGIS Pro. Click the About ArcGIS Pro button in the bottom left corner (denoted by the red arrow in the screenshot below). *In ArcGIS Pro v.2.3 this button is called Settings.* 
+
+![](Images\\manager1.jpg)
+
+2. In the menu on the left hand side click Python. This will open the Python Package Manager. Click the Manage Environments button.
+
+![](Images\\manager2a.jpg)
+
+3. This will open the Manage Environments window. Since additional packages cannot be installed in the default environment, you will need to clone the default environment. Ensure that the default environment is selected (this is typically called 'arcgispro-py3'), then click clone.  
+
+![](Images\\manager3.jpg)
+
+Creating the cloned environment may take a while, a blue line at the bottom of the window indicates that the process is still running. Once the clone has been created, click the radio button to make it the active environment. Click Close, then Exit ArcGIS Pro and restart it for the changes to take effect.
+
+![](Images\\manager3_1.jpg)
+
+
+4. Restart ArcGIS Pro, click on About ArcGIS Pro, then click on Python in the menu on the left hand side. Under Project Environment you should see the name of the cloned environment. *If you do not see the name of the cloned environment, repeat step 3 being careful not to exit the programme until the clone  has been created.* Click Add Packages. 
+
+![](Images\\manager4.jpg)
+
+This opens the Add Packages interface. In the Search box type patsy. The patsy package should appear in the list below. Select the patsy package and click Install. 
+
+![](Images\\manager4a.jpg)
+
+This will open the Install Package window. Tick the box in the bottom left to agree to the terms and conditions, then click Install. 
+
+![](Images\\manager4b.jpg)
+
+The installation may take a while. Once the installation is finished, the list underneath Add Packages will refresh. If you scroll down the list, you will see that patsy is no longer on it (because it has been installed). If you want to check that the package has been installed, click on Installed Packages. If you scroll down the list of installed packages, patsy should be listed.   
+
+5. Repeat step 4 for the following packages:
+
+- statsmodels
+
+![](Images\\manager4c.jpg)
+
+- seaborn 
+
+![](Images\\manager4d.jpg)
+
+- wxpython
+
+![](Images\\manager4e.jpg)
+
+After these four packages have been installed, click the back arrow in the menu on the left hand side. 
+
[…]
