### 2020-05-05 #2177 @anmolter
@Athelena I've added to the documentation that a Spatial Analyst license is required for analyses with raster data. I've also added an error pop up message to the code, if no license is found. While doing that I found two other bugs 😬  which should now be fixed 😜
