### 2020-05-05 #2177 @fortinma
@hugoledoux @anmolter and @Athelena, I apologize but I am unable to properly review this tool. I was able to finish the tutorial once, but I received different results than were expected. I think the different results were due mostly to my inattention because of how slowly my virtual machine was running ArcGIS Pro. After the initial run through, I tried to redo the tutorial with no avail as my virtual machines (all three that i tried) would not allow me to complete the tutorial as ArcGIS Pro kept crashing (I do have Spatial Analyst). I apologize again for not being able to complete the assignment.
