### 2020-05-06 #2177 @hugoledoux
Thanks for this @fortinma, the idea is indeed not that a review drives you mad 😆

The purpose of the JOSS review is to see if the software performs as claimed, if it can be installed and used by others. We aim to help the author(s) to achieve this, so you're report is certainly something to consider.

In this case, would you say that it's the VM that was the issue? Would ArcGIS Pro have crashed anyway, or is it due to the tool you reviewed? ArcGIS crashes a lot usually, hence me asking (sorry about that one, I couldn't help it 🤣).

If you could report briefly on this, that could help me decide how to proceed next with this review.

I know that @Athelena is running it too in a VM, do you have a similar experience?
