### 2020-05-06 #2177 @Athelena
I can't compare it to anything, I've only ever ran Pro on a VM, I don't think mine is necessarily that slow though. Right now my issue is that I don't have the right license and still waiting to hear back if I can get access. I'll let you know how it goes.
