### 2020-05-06 #2177 @fortinma
I have never really had any problems with ArcGIS Pro through a VM before and i wouldn't consider the usual running of it slow in the VM either. **However**, I never ran through high demand processing in it before. I wonder if **RAM** allocation to the VM might be an issue to consider?  Mine are set at **2 GB** and Esri recommends **8GB** for ArcGIS Pro.  Perhaps @Athelena can try with a higher RAM allocation to her VM when she gets Spatial Analyst installed?  I have a work computer with plenty of RAM, but running Windows 7. I can try again on that machine through Remote Desktop.
