### 2020-05-12 #2177 @fortinma
the one time I got through, that was also my issue. I did not have the predNO2 attribute
