### 2020-05-13 ce228af
Mistake in documentation, feature class with predicted values is called pred_lyr.
