### 2020-05-13 #2177 @anmolter
I've found the problem: I made a mistake in the documentation. The feature class with the predicted values is called pred_lyr not LURApply_receptors. If you open pred_lyr, there should be a predNO2 attribute. I've changed the documentation to show pred_lyr.
