### 2020-05-13 #2177 @Athelena
I am not an expert in LUR modelling but as this software is aimed at GIS specialist I’m responding from that perspective.

My notes on the review based on the review checklist:

Documentation
-Functionality documentation: While the set up, over-view, and tutorial are well documented there is very little said about the LUR model itself. What are the core principles? How does the multiple linear regression analysis actually work? On that note there should be a section that helps your interpret the outputs, I see in the output folder many matrices, box plots, etc., but you wouldn’t know this from the guide. If you are targeting GIS specialist it would be helpful to not assume only domain specialist would use this, perhaps this can be adapted for other applications as well. So for me there needs to be a clear understanding of what everything in the output is. Also I see that the pdf document has more information than the html pages, so it’s not clear which guide to follow.
-Community guidelines: It would be nice if the guides linked back to the repo and if the repo encouraged contribution from others.

Software paper
-State of the field: I don’t see an overview of how this compares to other software. I also have one off the top of the head, but I haven’t used it in years so forgive me if not relevant: https://naturalcapitalproject.stanford.edu/software/invest
-References: I don’t know the sources for the sample data or the python packages utilised, and even a link to ArcGIS Pro is probably useful.

Over-all following the guide, in the end it was relatively straight-forward but the guide ends at execution so it’s not clear what to do after. Also small notes, the first time I tried to run Apply after Build it resulted in a crash but after that it ran fine, don’t know if it was a coincidence or maybe this might happen to others and the tools can’t be run in the same session, just flagging it just in case. Also I ran Apply twice and my results match for the pred_lyr for both times but they don’t match to your results in the screenshot, so I’m not sure if it’s an issue of different data or of different results in the statistical modelling, which is why understanding the outputs is crucial.
