### 2020-05-15 #2177 @anmolter
Thanks @Athelena , that's very helpful. I'll try to make changes over the next week or so. Just a quick question for clarification: you said there should be more information on the output files that are created, do you mean in the tutorial section or in the help section for the output folder (p19 in pdf)?
