### 2020-05-15 #2177 @Athelena
Mainly for the tutorial section, I was following the html version. Also a link to the ESCAPE methodology would be useful, I tried to go to their website but it seems to be down? It's hard to interpret the results otherwise.
