### 2020-05-20 e267db7
Added more information to documentation.
