+++ XLUR/Code/Documentation/XLUR_documentation.Rmd
@@ -2,11 +2,14 @@
 title: XLUR Wizard
 date: "v.`r Sys.Date()`"
 output:
-  output: rmarkdown::github_document
+#   rmdformats::material:
+#     highlight: kate
+#     self_contained: no
+#     lightbox: TRUE
+#     fig_caption: FALSE
+ output: rmarkdown::github_document
 
 ---
-
-
 
 ```{r knitr_init, echo=FALSE, cache=FALSE}
 library(knitr)
@@ -24,6 +27,7 @@ opts_chunk$set(echo=FALSE,
                fig.path = "documentation_figs/README-")
 opts_knit$set(width=75)
 ```
+
 
 # Installation
 
@@ -82,15 +86,30 @@ After these four packages have been installed, click the back arrow in the menu 
 
 # General Information {.tabset .tabset-pills}
 
-## What is LUR?
+## What is LUR? 
+
+**Classic LUR**
 
 Land use regression (LUR) is a statistical method, which uses geospatial data to develop prediction models in environmental sciences. It is predominantly used in air pollution research to predict pollutant concentrations empirically within a given a study area. However, it has also been used for other environmental phenomena such as noise, air temperature and water microbiology.   
 
 The underlying principle of LUR is that a measured quantity (e.g. pollutant concentration, noise level, temperature etc) at a given location depends on characteristics of the surrounding environment, in particular on the presence and absence of  sources and sinks, which increase and decrease values respectively. 
 
-LUR models are developed by using measured data from a number of monitoring sites as the dependent variable and data on the surrounding environment extracted as predictor variables in a multiple linear regression analysis. For example, in an air pollution study particulate matter concentrations might be measured at fifty monitoring sites. Then for each monitoring site potential predictor variables are extracted such as the area of industrial land use around the monitoring site, the distance to the nearest road, the number of motor vehicles on the nearest road etc. The particulate matter concentrations and the potential predictors are entered into a supervised machine learning process, which will try to construct a parsimonious multiple linear regression model. This model can then be used to predict particulate matter concentrations at any point within the study area.
+LUR models are developed by using measured data from a number of monitoring sites as the dependent variable and data on the surrounding environment extracted as potential predictor variables in a multiple linear regression analysis. For example, in an air pollution study particulate matter concentrations might be measured at fifty monitoring sites. Then for each monitoring site potential predictor variables are extracted such as the area of industrial land use around the monitoring site, the distance to the nearest road, the number of motor vehicles on the nearest road etc. The particulate matter concentrations and the potential predictors are entered into a supervised machine learning process, which will try to construct a parsimonious multiple linear regression model. This model can then be used to predict particulate matter concentrations at any point within the study area.
 
-Hybrid models can also be developed. In the XLUR tool a hybrid model is a model in which one or more mandatory variables are selected by the user. These mandatory variables are forced into the model, regardless of the amount of variance that they explain or their direction of effect. An example of this would be using the output from dispersion models run at a coarse spatial resolution as a mandatory variable, which could add a measure of global variability to the measures of local variability used in LUR.
+The supervised machine learning process is based on the methodology used in the European Study of Cohorts for Air Pollution Effects (ESCAPE), which can be downloaded from [http://www.escapeproject.eu/manuals/ESCAPE_Exposure-manualv9.pdf](http://www.escapeproject.eu/manuals/ESCAPE_Exposure-manualv9.pdf) (a copy can be found in the Documentation folder). The ESCAPE exposure manual provides a detailed description of the steps required to construct a parsimonious multiple linear regression model; therefore, only a brief summary is presented here:
+
+1. Prior to the statistical analysis a positive or negative direction of effect is assigned to each potential predictor variable by the user based on *a priori* knowledge of the subject area. 
[…]
