+++ README.md
@@ -4,6 +4,6 @@ A Python toolbox for ArcGIS Pro that enables the development and application of 
 ## Instructions
 XLUR was specifically developed as a toolbox for use within the ArcGIS Pro (v2.2.4 or higher, Environmental Systems Research Institute (ESRI), Redlands, CA) software. To use XLUR the following steps are recommended:
 1. Clone or download the XLUR repository to a local drive. Avoid having spaces in the file path to the directory where you are storing the XLUR repository. 
-2. In the XLUR repository open the XLUR folder and doubleclick XLUR_Manual_BrowserFriendly - Shortcut. This will open the XLUR Wizard manual in a standard web browser. The manual contains the documentation for the XLUR tools and a tutorial section, which guides the user through a test run of XLUR. Alternatively, you can <a href="https://anmolter.github.io/XLUR/XLUR_documentation.pdf" type="application/pdf"> click here </a> to open a pdf version of the user manual in your web browser.
+2. In the XLUR repository open the XLUR folder and doubleclick XLUR_documentation_Shortcut. This will open the XLUR Wizard manual in a standard web browser. The manual contains the documentation for the XLUR tools and a tutorial section, which guides the user through a test run of XLUR. Alternatively, you can <a href="https://anmolter.github.io/XLUR/XLUR_documentation.pdf" type="application/pdf"> click here </a> to open a pdf version of the user manual in your web browser.
 3. Follow the instructions in the manual to install XLUR and to launch the tools.  
 
