### 2020-05-21 c65be4c
Added state of the field and availability sections.
