+++ paper.bib
@@ -1,4 +1,10 @@
-﻿@article{Beelen:2013,
+﻿@misc{ArcGISPro,
+   url = {https://pro.arcgis.com/en/pro-app/get-started/get-started.htm},
+   year = {2020},
+   type = {Web Page}
+}
+
+@article{Beelen:2013,
    author = {Beelen, R. and Hoek, G. and Vienneau, D. and Eeftens, M. and Dimakopoulou, K. and Pedeli, X. and Tsai, M. Y. and Kunzli, N. and Schikowski, T. and Marcon, A. and Eriksen, K. T. and Raaschou-Nielsen, O. and Stephanou, E. and Patelarou, E. and Lanki, T. and Yli-Tuomi, T. and Declercq, C. and Falq, G. and Stempfelet, M. and Birk, M. and Cyrys, J. and von Klot, S. and Nador, G. and Varro, M. J. and Dedele, A. and Grazuleviciene, R. and Molter, A. and Lindley, S. and Madsen, C. and Cesaroni, G. and Ranzi, A. and Badaloni, C. and Hoffmann, B. and Nonnemacher, M. and Kraemer, U. and Kuhlbusch, T. and Cirach, M. and de Nazelle, A. and Nieuwenhuijsen, M. and Bellander, T. and Korek, M. and Olsson, D. and Stromgren, M. and Dons, E. and Jerrett, M. and Fischer, P. and Wang, M. and Brunekreef, B. and de Hoogh, K.},
    title = {Development of NO2 and NOx land use regression models for estimating air pollution exposure in 36 study areas in Europe - The ESCAPE project},
    journal = {Atmospheric Environment},
@@ -73,6 +79,33 @@
    type = {Journal Article}
 }
 
+@article{Lautenschlager:2020,
+   author = {Lautenschlager, Florian and Becker, Martin and Kobs, Konstantin and Steininger, Michael and Davidson, Padraig and Krause, Anna and Hotho, Andreas},
+   title = {OpenLUR: Off-the-shelf air pollution modeling with open features and machine learning},
+   journal = {Atmospheric Environment},
+   volume = {233},
+   pages = {117535},
+   ISSN = {1352-2310},
+   DOI = {https://doi.org/10.1016/j.atmosenv.2020.117535},
+   url = {http://www.sciencedirect.com/science/article/pii/S1352231020302703},
+   year = {2020},
+   type = {Journal Article}
+}
+
+@article{Ma:2020,
+   author = {Ma, Xuying and Longley, Ian and Salmond, Jennifer and Gao, Jay},
+   title = {PyLUR: Efficient software for land use regression modeling the spatial distribution of air pollutants using GDAL/OGR library in Python},
+   journal = {Frontiers of Environmental Science & Engineering},
+   volume = {14},
+   number = {3},
+   pages = {44},
+   ISSN = {2095-221X},
+   DOI = {10.1007/s11783-020-1221-5},
+   url = {https://doi.org/10.1007/s11783-020-1221-5},
+   year = {2020},
+   type = {Journal Article}
+}
+
 @article{Molter:2010a,
    author = {Molter, A. and Lindley, S. and de Vocht, F. and Simpson, A. and Agius, R.},
    title = {Modelling air pollution for epidemiologic research--Part I: A novel approach combining land use regression and air dispersion},
@@ -103,3 +136,15 @@
    type = {Journal Article}
 }
 
+@article{Morley:2018,
+   author = {Morley, David W. and Gulliver, John},
+   title = {A land use regression variable generation, modelling and prediction tool for air pollution exposure assessment},
+   journal = {Environmental Modelling & Software},
+   volume = {105},
+   pages = {17-23},
[…]
