+++ Packages.md
@@ -1,0 +1,33 @@
+# Python packages
+
+XLUR uses a number of Python modules/packages. Most of these are Python base modules or are pre-installed with
+ArcGIS Pro. Only four additional packages need to be installed via the Python Package Manager. These packages are 
+required to create the GUI, to run statistical analyses, and to create some of the output graphs. The following
+modules/packages are required: 
+
+## Python base modules (https://docs.python.org/3/library/)
+* sys
+* os
+* logging
+* sqlite3
+* csv
+* time
+* itertools
+* math
+* string
+* random
+* shutil
+* webbrowser
+
+## Python packages pre-installed with ArcGIS Pro
+* arcpy (https://pro.arcgis.com/en/pro-app/arcpy/get-started/what-is-arcpy-.htm)
+* pandas (https://pandas.pydata.org/)
+* numpy (https://numpy.scipy.org/)
+* matplotlib (https://matplotlib.org/)
+
+
+## Installation required via Python Package Manager
+* seaborn (http://seaborn.pydata.org/)
+* patsy (https://github.com/pydata/patsy)
+* statsmodels (https://www.statsmodels.org/stable/index.html)
+* wx (https://wxpython.org/)
