### 2020-05-22 #2177 @anmolter
Based on the comments by @athelena I have made the following changes:

- I have added more information to the manual in the General Information > What is LUR section on how the regression analysis works and on how the hybrid model works. I have also added a copy of the ESCAPE exposure manual pdf to the documentation folder in case their website is down.
- I have added more information to the manual in the Tutorial section explaining the outputs from the Build LUR tool and the Apply LUR tool.
- The pdf manual and the html manual are now outputs from the same markdown files (XLUR_documentation.rmd), so their content should now be identical.
- I have added a Contributions.md to the repository and I have added a link to the repository to the paper. In addition, there are now some links in the manual to specific sections of the repository.
- I have added a section to the paper on the State of the field. I don’t think the InVEST tool is a LUR (although it looks quite interesting), but there are three other tools that I know of. So I have explained what they are and how they differ to XLUR.
- I have added Packages.md to the repository, which lists all of the Python packages and provides source links for the non-base packages. I have also added a reference to ArcGIS Pro to the paper, and I have added a link to the ExampleData_Sources.md to the tutorial section in the manual.
- I have added more information to the tutorial section on what can be done after running the Apply tool.
- ArcGIS crashing when the tool runs more than once is a known issue  :frowning_face: (it’s listed in Troubleshooting). I think it has something to do with the warning message about the threads. We have reported this to ESRI, but I don’t know, if they’re working on a fix for this :disappointed:  I have added to the tutorial section that the user should restart ArcGIS before running the Apply tool, so then it shouldn’t crash.
- I have added a folder called TutorialOutput to the repository, which shows the output files that I get when I follow the tutorial. I ran it twice and I got the same result both times. But please let me know, if you get something else with the same inputs! I have not included the geodatabases and sqlite databases in this folder to avoid it being too large, but the LOG files and the out_pred.csv should allow a user to compare their results to mine.

Thank you again for your comments, I am quite new to this, so they were very helpful.
