### 2020-05-25 #2177 @Athelena
I am now satisfied, all of my comments were sufficiently addressed. I completed the checklist above.

This was a new and interesting experience for me as well!
