### 2020-06-04 #2177 @fortinma
Good morning,
I have made another four attempts, and the software continues to hang when selecting values for the raster cell (step G from the Predictors Page).  The process is running and doesn't hang, but doesn't seem to want to stop. I ran the last one for 12 hours just to make sure, and then had to kill it to get out.  I also checked the task manager and the task is still running and not stopped. I am running it again now, but it has been running for about 45 minutes.

![image](https://user-images.githubusercontent.com/1124016/83738851-d3290080-a622-11ea-8d84-a527e7e15c0b.png)
![image](https://user-images.githubusercontent.com/1124016/83738896-dfad5900-a622-11ea-9f74-29e6da617359.png)
