### 2020-06-04 #2177 @Athelena
Maybe a silly question but I had this when I didn't have Spatial Analyst enabled, just to double check?
