### 2020-06-04 #2177 @fortinma
not a silly question -- silly reviewer did not have Spatial Analyst on! thanks. Will try again.
