### 2020-06-04 #2177 @Athelena
I know Spatial Analyst is listed as a requirement but I think it should be flagged in the script at the beginning or you can automatically check it out I believe:

https://gis.stackexchange.com/questions/151166/accessing-spatial-analyst-extension-from-stand-alone-script
