### 2020-06-04 #2177 @fortinma
It might also be useful to add to the tutorial to check that the SA license is available in the licensing tab.  This is especially true because  named user licenses do not include both ArcGIS Pro and its extensions by default (at least at most institutions).  Administrators need to turn on the extensions.
