### 2020-06-04 #2177 @fortinma
I have completed the checklist and am satisfied with the results of running the tutorial. All seems to be in order. Spatial Analyst being the common issue in both reviews, through no fault of the author but because of the nature of access to the extension through the esri portal, a popup stating Spatial Analyst is required if not available in the users installation.  Alternatively, a simple statement in the tutorial on making sure the extension license is available (it is not simply turned on within the software as it was in ArcMap previously) would be welcome. Neither option, however, should be a requirement for moving this review forward. Thank you. Marcel
