### 2020-06-04 #2177 @hugoledoux
OK @anmolter, the 2 reviewers are now happy, so we're reaching the end of the review!

At this point could you:
- [ ] Add what the 2 reviewers pointed out above with Spatial Analyst
- [ ] Make a tagged release of your software (you have none yet), and list the version tag of the archived version here.
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
