### 2020-06-04 0da1c8b
Added reminder to check Spatial Analyst license to documentation.
