+++ XLUR/Code/Documentation/XLUR_documentation.Rmd
@@ -80,9 +80,9 @@ The installation may take a while. Once the installation is finished, the list u
 
 ![](Images\\manager4e.jpg)
 
-After these four packages have been installed, click the back arrow in the menu on the left hand side. 
+6. If you are planning to use raster data in your analysis, click on Licensing in the menu on the left hand side. Under Esri Extensions check that Spatial Analyst is licensed.  
 
-6. Click Open another project, browse to the XLUR.aprx ArcGIS Pro Project file and double-click to open it. The XLUR.aprx file can be found in the XLUR folder in the XLUR repository. In the Catalog window double-click Toolboxes, then double-click XLUR.tbx. This will open the XLUR toolbox, which contains the BuildLUR and ApplyLUR scripts. Running either of these script will open the Build LUR or Apply LUR wizard, respectively.
+7.  Click the back arrow in the menu on the left hand side. Click Open another project, browse to the XLUR.aprx ArcGIS Pro Project file and double-click to open it. The XLUR.aprx file can be found in the XLUR folder in the XLUR repository. In the Catalog window double-click Toolboxes, then double-click XLUR.tbx. This will open the XLUR toolbox, which contains the BuildLUR and ApplyLUR scripts. Running either of these script will open the Build LUR or Apply LUR wizard, respectively.
 
 # General Information {.tabset .tabset-pills}
 
@@ -811,7 +811,7 @@ htmltools::includeHTML("p3F_SetSourceSink.html")
 
 #### Value of Raster cell {.tabset .tabset-pills}
 
-For this type of variable a raster grid file should be used, which ideally has a spatial extent that is larger than the study area. The wizard will extract the value of the raster cell that is spatially coincident with the point location representing the monitoring site (dependent variable). 
+You must have a **Spatial Analyst license** to create this type of predictor. For this type of variable a raster grid file should be used, which ideally has a spatial extent that is larger than the study area. The wizard will extract the value of the raster cell that is spatially coincident with the point location representing the monitoring site (dependent variable). 
 
 An example of the use of this predictor variable type is elevation. Elevation is commonly sourced from a Digital Elevation Model stored as a raster grid. 
 
@@ -987,6 +987,10 @@ Enlarge the window by clicking the maximise button in the title bar.
 
 Please inspect your receptor points carefully on a map and look at the predictor variables in your LUR models. If the LUR models contain predictors that are derived from the Inverse distance, Inverse distance squared, Value \* Inverse Distance, or Value \* Inverse distance squared to the nearest feature method, then any receptor points spatially coincident with the nearest feature will be dropped to prevent a division by zero error.
 
+* ***The Build LUR tool freezes when I use the Value of Raster cell page***
+
+Please check that you have a license for Spatial Analyst. 
+
 # Tutorial
 
 This section provides a guided tutorial for building and applying a LUR model using the XLUR wizard. In this tutorial you will build and apply a predictive air pollution model for the Greater Manchester area using [openly accessible datasets](https://github.com/anmolter/XLUR/blob/master/ExampleData/ExampleData_Sources.md) on monitored Nitrogen Dioxide (NO~2~) concentrations, land use categories, road networks, and emission sources. Please note that the purpose of this tutorial is to illustrate the use of the XLUR wizard, not to develop a high performing LUR model; therefore, only a small number of input datasets are used in this tutorial.   
@@ -994,7 +998,7 @@ This section provides a guided tutorial for building and applying a LUR model us
 ## Before you start
 
 1. Ensure that you have cloned or downloaded the XLUR repository from GitHub, including the ExampleData folder.
-2. Follow the instructions in the [Installation](#installation) section to ensure that the additional Python packages required by XLUR have been installed.
[…]
