### 2020-06-04 #2177 @anmolter
Hi @hugoledoux , I've added several reminders to the documentation (Installation, FAQs, Tutorial) that a Spatial Analyst license is required and explained in Installation how to check that the license is there. I've done a tagged release, I've archived the software in Zenodo and I've sorted out the metadata. The DOI is 10.5281/zenodo.3877328. I hope everything is OK now.
