### 2020-06-05 #2177 @hugoledoux
Thanks.

One last request if possible: that the code pops up a warning/alert is the extension is not available? Is that something possible?
