### 2020-06-05 #2177 @Athelena
Yes you can check it out at the beginning:

https://pro.arcgis.com/en/pro-app/help/analysis/spatial-analyst/mapalgebra/importing-the-spatial-analyst-module.htm
