### 2020-06-05 a0a1a53
Changed error catching for spatial analyst license error
