### 2020-06-05 #2177 @anmolter
Originally, I had the checkOutExtension code at the beginning, but after @Athelena had problems I moved it to the Raster predictor page (so that someone without a license could still use the vector parts of the tool) and I added error catching with a pop up message. But it looks like that didn't work, because @fortinma didn't get that pop up 😞 I've amended the code now based on this link https://pro.arcgis.com/en/pro-app/arcpy/functions/checkoutextension.htm . Unfortunately I can't test if the message appears, because I'm on a named user license.
