### 2020-06-05 #2177 @Athelena
Well I think the checking out should be at the beginning. I think that good practice for arcpy is to enable extensions directly after importing it (so near the beginning of your script). I think if you want parts of the tool to be usable for people without a license then I think it should be a separate tool in the toolbox or you flag the issue earlier.

I also can't test in Pro but I did some crude tests of just the code and I can see that your current code will fail because `LicenseError` is not defined. I believe you need to have before your try statement this:

```
class LicenseError(Exception):
    pass
```

This is the same as in the link you provided.

I've also never used `wx` so not sure how that will impact the error message.
