### 2020-06-05 d5b5c2c
Fixed problem with Spatial Analyst error message.
