### 2020-06-05 #2177 @anmolter
I got rid of the try-except bit and I'm just using if-else now, which makes it much easier. I also managed to test it now, by temporarily setting it to check for no license (I probably should've thought of that earlier) and it finally works. I agree it would be better to have the license check code at the beginning, but when I tried that it didn't work. I think because the wxApp hadn't been created yet. However, the way it is set up now the error pop up message will appear straight away, even before the first page of the wizard appears.
