### 2020-06-05 #2177 @fortinma
I can take away my Spatial Analyst extension temporarily to test it if you want?
