### 2020-06-08 #2177 @hugoledoux
🎉

One last thing needs to be done @anmolter : you need to fix the 2 DOIs as pointed just above. You need to remove the `http://dx.doi.org/` from the "DOI" fields for 2 references in your BIB, and I guess generate a new version (or erase 1.0.0 put that one, it's possible I reckon).
