+++ paper.bib
@@ -86,7 +86,7 @@
    volume = {233},
    pages = {117535},
    ISSN = {1352-2310},
-   DOI = {https://doi.org/10.1016/j.atmosenv.2020.117535},
+   DOI = {10.1016/j.atmosenv.2020.117535},
    url = {http://www.sciencedirect.com/science/article/pii/S1352231020302703},
    year = {2020},
    type = {Journal Article}
@@ -143,7 +143,7 @@
    volume = {105},
    pages = {17-23},
    ISSN = {1364-8152},
-   DOI = {https://doi.org/10.1016/j.envsoft.2018.03.030},
+   DOI = {10.1016/j.envsoft.2018.03.030},
    url = {http://www.sciencedirect.com/science/article/pii/S1364815217302621},
    year = {2018},
    type = {Journal Article}
