### 2020-06-08 #2177 @anmolter
@hugoledoux Do you mean generate a new release? Will that mess with the Zenodo link?
