### 2020-06-11 #2177 @hugoledoux
Yeah, it will. It's a small thing, but still could you generate v1.0.1, zenodo it, and give me the DOI here?
