### 2020-06-11 #2177 @anmolter
Done. The new DOI is 10.5281/zenodo.3889604
![image](https://user-images.githubusercontent.com/34544873/84357868-389a6580-abbe-11ea-9725-44344c4f3852.png)
