### 2020-06-11 #2177 @labarba
@anmolter -- please edit the metadata of your Zenodo deposit (title and author fields) to match the JOSS paper.

In the paper, I found this problematic sentence:
> Unlike XLUR, RLUR and PyLUR it does not use the ESCAPE methodology…

It is unclear what the subject of the sentence is: "it does not use..." makes the subject "it" but before that you have three names of software, and we don't know which of these is the one that does not use the ESCAPE methodology. Please fix.

Also, it is kind of verbose to have "Journal article" as an added string with every cited publication. Please remove that.
