+++ paper.bib
@@ -13,8 +13,7 @@
    ISSN = {1352-2310},
    DOI = {10.1016/j.atmosenv.2013.02.037},
    url = {<Go to ISI>://WOS:000318262000002},
-   year = {2013},
-   type = {Journal Article}
+   year = {2013}
 }
 
 @article{deHoogh:2016,
@@ -26,8 +25,7 @@
    ISSN = {0013-9351},
    DOI = {10.1016/j.envres.2016.07.005},
    url = {http://www.sciencedirect.com/science/article/pii/S0013935116302894},
-   year = {2016},
-   type = {Journal Article}
+   year = {2016}
 }
 
 @article{Eeftens:2012,
@@ -41,8 +39,7 @@
 0013-936X (Linking)},
    DOI = {10.1021/es301948k},
    url = {http://www.ncbi.nlm.nih.gov/pubmed/22963366},
-   year = {2012},
-   type = {Journal Article}
+   year = {2012}
 }
 
 @misc{escape,
@@ -61,8 +58,7 @@
    ISSN = {2169-897x},
    DOI = {10.1002/2012jd019399},
    url = {<Go to ISI>://WOS:000338338100009},
-   year = {2014},
-   type = {Journal Article}
+   year = {2014}
 }
 
 @article{Kelsey:2004,
@@ -75,8 +71,7 @@
    ISSN = {0022-0981},
    DOI = {10.1016/S0022-0981(03)00359-9},
    url = {http://www.sciencedirect.com/science/article/pii/S0022098103003599},
-   year = {2004},
-   type = {Journal Article}
+   year = {2004}
 }
 
 @article{Lautenschlager:2020,
@@ -88,8 +83,7 @@
    ISSN = {1352-2310},
    DOI = {10.1016/j.atmosenv.2020.117535},
    url = {http://www.sciencedirect.com/science/article/pii/S1352231020302703},
-   year = {2020},
-   type = {Journal Article}
+   year = {2020}
 }
 
 @article{Ma:2020,
@@ -102,8 +96,7 @@
    ISSN = {2095-221X},
    DOI = {10.1007/s11783-020-1221-5},
    url = {https://doi.org/10.1007/s11783-020-1221-5},
-   year = {2020},
-   type = {Journal Article}
+   year = {2020}
 }
 
 @article{Molter:2010a,
@@ -117,8 +110,7 @@
 0048-9697 (Linking)},
    DOI = {10.1016/j.scitotenv.2010.08.027},
    url = {http://www.ncbi.nlm.nih.gov/pubmed/20846708},
-   year = {2010},
-   type = {Journal Article}
+   year = {2010}
 }
 
 @article{Molter:2010b,
@@ -132,8 +124,7 @@
 0048-9697 (Linking)},
    DOI = {10.1016/j.scitotenv.2010.10.005},
    url = {http://www.ncbi.nlm.nih.gov/pubmed/20970170},
-   year = {2010},
-   type = {Journal Article}
+   year = {2010}
 }
 
 @article{Morley:2018,
@@ -145,6 +136,5 @@
    ISSN = {1364-8152},
    DOI = {10.1016/j.envsoft.2018.03.030},
    url = {http://www.sciencedirect.com/science/article/pii/S1364815217302621},
-   year = {2018},
-   type = {Journal Article}
+   year = {2018}
 }
