### 2020-06-11 #2177 @anmolter
@labarba I've made the changes. Zenodo reset everything to default when I did a new release, but I've edited the metadata now, so everything should be correct.
